LayoutTests/http/tests/websocket/tests/hybi/reserved-opcodes-expected.txt

CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 3
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 4
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 5
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 6
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 7
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 11
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 12
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 13
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 14
CONSOLE MESSAGE: line 0: Unrecognized frame opcode: 15
Tests whether WebSocket ignores frames with reserved opcode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Opcode 3: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 4: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 5: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 6: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 7: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 11: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 12: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 13: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 14: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
Opcode 15: Test started.
onopen() was called.
onclose() was called.
PASS closeEvent.wasClean is false
PASS successfullyParsed is true

TEST COMPLETE
